requirement.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.1.0
bitarray==3.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
ckzg==2.0.1
cytoolz==1.0.1
eth-account==0.13.5
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.2.0
eth-typing==5.2.0
eth-utils==5.2.0
eth_abi==5.0.1
frozenlist==1.5.0
hexbytes==1.3.0
idna==3.10
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lru-dict==1.3.0
multidict==6.1.0
oauthlib==3.2.2
parsimonious==0.9.0
propcache==0.3.0
protobuf==5.29.3
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
pyunormalize==16.0.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rlp==4.1.0
rpds-py==0.23.1
toolz==1.0.0
tweepy==4.15.0
types-requests==2.32.0.20241016
typing_extensions==4.12.2
urllib3==2.3.0
web3==6.0.0
websockets==13.1
yarl==1.18.3
